Add files via upload
SFO Public Department, referred to as SFO, has captured the salary data of all its employees for the years 2011-2014. Now in 2018, the organization is facing some financial crisis. As a first step, HR wants to rationalize employee cost to save payroll budget. You have to perform data manipulation and answer the questions below:

1. How much total salary cost has increased from year 2011 to 2014?
2. Who was the top-earning employee across all the years?